refactor(tests): refactored out HttpClientFactory
