+++ src/FrontierSharp.Tests/FrontierSharp.Tests.csproj
@@ -19,6 +19,7 @@
       </PackageReference>
       <PackageReference Include="AwesomeAssertions.Json" Version="8.0.0" />
       <PackageReference Include="FluentResults" Version="3.16.0" />
+      <PackageReference Include="Microsoft.Extensions.Caching.Hybrid" Version="9.3.0" />
       <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.13.0" />
       <PackageReference Include="NSubstitute" Version="5.3.0" />
       <PackageReference Include="NSubstitute.Analyzers.CSharp" Version="1.0.17">
