+++ src/FrontierSharp.Tests/Utils/FakeHybridCache/FakeHybridCache.cs
@@ -1,0 +1,23 @@
+using Microsoft.Extensions.Caching.Hybrid;
+
+namespace FrontierSharp.Tests.Utils.FakeHybridCache;
+
+internal sealed class FakeHybridCache : HybridCache {
+    public override ValueTask<T> GetOrCreateAsync<TState, T>(string key, TState state, Func<TState, CancellationToken, ValueTask<T>> factory,
+        HybridCacheEntryOptions? options = null, IEnumerable<string>? tags = null, CancellationToken cancellationToken = default) {
+        return factory(state, cancellationToken);
+    }
+
+    public override ValueTask RemoveAsync(string key, CancellationToken cancellationToken = default) {
+        return default;
+    }
+
+    public override ValueTask RemoveByTagAsync(string tag, CancellationToken cancellationToken = default) {
+        return default;
+    }
+
+    public override ValueTask SetAsync<T>(string key, T value, HybridCacheEntryOptions? options = null, IEnumerable<string>? tags = null,
+        CancellationToken cancellationToken = default) {
+        return default;
+    }
+}
